+++ CHANGELOG.md
@@ -85,6 +85,7 @@
    * CHANGED: Removed unused headers [#5254](https://github.com/valhalla/valhalla/pull/5254)
    * ADDED: "destination_only_hgv" in directed edge json [#5281](https://github.com/valhalla/valhalla/pull/5281)
    * CHANGED: Link libvalhalla to libgeos. Build command to use `nmake` on Windows instead of `make`. Skipping check for `CMAKE_BUILD_TYPE` when using a multi-config generator like Visual Studio or XCode. [#5294](https://github.com/valhalla/valhalla/pull/5294)
+   * ADDED: workflow to publish Python bindings for all major platforms to PyPI [#5280](https://github.com/valhalla/valhalla/pull/5280)
 
 ## Release Date: 2024-10-10 Valhalla 3.5.1
 * **Removed**